per #344, moved rest of functions from met_util into more appropriate util files and removed met_util, ci-run-all-diff
