+++ metplus/util/time_looping.py
@@ -1,9 +1,12 @@
+import re
 from datetime import datetime, timedelta
 
-from .string_manip import getlist
-from .time_util import get_relativedelta
+from .string_manip import getlist, getlistint
+from .time_util import get_relativedelta, add_to_time_input
+from .time_util import ti_get_hours_from_relativedelta
+from .time_util import ti_get_seconds_from_relativedelta
 from .string_template_substitution import do_string_sub
-
+from .config_metplus import log_runtime_banner
 
 def time_generator(config):
     """! Generator used to read METplusConfig variables for time looping
@@ -121,6 +124,41 @@ def get_start_and_end_times(config):
         return None, None
 
     return start_dt, end_dt
+
+
+def loop_over_times_and_call(config, processes, custom=None):
+    """! Loop over all run times and call wrappers listed in config
+
+    @param config METplusConfig object
+    @param processes list of CommandBuilder subclass objects (Wrappers) to call
+    @param custom (optional) custom loop string value
+    @returns list of tuples with all commands run and the environment variables
+    that were set for each
+    """
+    # keep track of commands that were run
+    all_commands = []
+    for time_input in time_generator(config):
+        if not isinstance(processes, list):
+            processes = [processes]
+
+        for process in processes:
+            # if time could not be read, increment errors for each process
+            if time_input is None:
+                process.errors += 1
+                continue
+
+            log_runtime_banner(config, time_input, process)
+            add_to_time_input(time_input,
+                              instance=process.instance,
+                              custom=custom)
+
+            process.clear()
+            process.run_at_time(time_input)
+            if process.all_commands:
+                all_commands.extend(process.all_commands)
+            process.all_commands.clear()
+
+    return all_commands
 
 
 def _validate_time_values(start_dt, end_dt, time_interval, prefix, logger):
@@ -287,3 +325,211 @@ def skip_time(time_info, skip_times):
 
     # if skip time never matches, return False
     return False
+
+
+def get_lead_sequence(config, input_dict=None, wildcard_if_empty=False):
+    """!Get forecast lead list from LEAD_SEQ or compute it from INIT_SEQ.
+        Restrict list by LEAD_SEQ_[MIN/MAX] if set. Now returns list of relativedelta objects
+        Args:
+            @param config METplusConfig object to query config variable values
+            @param input_dict time dictionary needed to handle using INIT_SEQ. Must contain
+               valid key if processing INIT_SEQ
+            @param wildcard_if_empty if no lead sequence was set, return a
+             list with '*' if this is True, otherwise return a list with 0
+            @returns list of relativedelta objects or a list containing 0 if none are found
+    """
+
+    out_leads = []
+    lead_min, lead_max, no_max = _get_lead_min_max(config)
+
+    # check if LEAD_SEQ, INIT_SEQ, or LEAD_SEQ_<n> are set
+    # if more than one is set, report an error and exit
+    lead_seq = getlist(config.getstr('config', 'LEAD_SEQ', ''))
+    init_seq = getlistint(config.getstr('config', 'INIT_SEQ', ''))
+    lead_groups = get_lead_sequence_groups(config)
+
+    if not _are_lead_configs_ok(lead_seq,
+                                init_seq,
+                                lead_groups,
+                                config,
+                                input_dict,
+                                no_max):
+        return None
+
+    if lead_seq:
+        # return lead sequence if wildcard characters are used
+        if lead_seq == ['*']:
+            return lead_seq
+
+        out_leads = _handle_lead_seq(config,
+                                    lead_seq,
+                                    lead_min,
+                                    lead_max)
+
+    # use INIT_SEQ to build lead list based on the valid time
+    elif init_seq:
+        out_leads = _handle_init_seq(init_seq,
+                                    input_dict,
+                                    lead_min,
+                                    lead_max)
+    elif lead_groups:
+        out_leads = _handle_lead_groups(lead_groups)
+
+    if not out_leads:
+        if wildcard_if_empty:
+            return ['*']
+
+        return [0]
+
+    return out_leads
+
+def _are_lead_configs_ok(lead_seq, init_seq, lead_groups,
+                         config, input_dict, no_max):
+    if lead_groups is None:
+        return False
+
+    error_message = ('are both listed in the configuration. '
+                     'Only one may be used at a time.')
+    if lead_seq:
+        if init_seq:
+            config.logger.error(f'LEAD_SEQ and INIT_SEQ {error_message}')
+            return False
+
+        if lead_groups:
+            config.logger.error(f'LEAD_SEQ and LEAD_SEQ_<n> {error_message}')
+            return False
+
+    if init_seq and lead_groups:
+        config.logger.error(f'INIT_SEQ and LEAD_SEQ_<n> {error_message}')
+        return False
+
+    if init_seq:
+        # if input dictionary not passed in,
+        # cannot compute lead sequence from it, so exit
+        if input_dict is None:
+            config.logger.error('Cannot run using INIT_SEQ for this wrapper')
+            return False
+
+        # if looping by init, fail and exit
+        if 'valid' not in input_dict.keys():
+            log_msg = ('INIT_SEQ specified while looping by init time.'
+                       ' Use LEAD_SEQ or change to loop by valid time')
+            config.logger.error(log_msg)
+            return False
+
+        # maximum lead must be specified to run with INIT_SEQ
+        if no_max:
+            config.logger.error('LEAD_SEQ_MAX must be set to use INIT_SEQ')
+            return False
+
+    return True
+
+def _get_lead_min_max(config):
+    # remove any items that are outside of the range specified
+    #  by LEAD_SEQ_MIN and LEAD_SEQ_MAX
+    # convert min and max to relativedelta objects, then use current time
+    # to compare them to each forecast lead
+    # this is an approximation because relative time offsets depend on
+    # each runtime
+    huge_max = '4000Y'
+    lead_min_str = config.getstr_nocheck('config', 'LEAD_SEQ_MIN', '0')
+    lead_max_str = config.getstr_nocheck('config', 'LEAD_SEQ_MAX', huge_max)
+    no_max = lead_max_str == huge_max
+    lead_min = get_relativedelta(lead_min_str, 'H')
+    lead_max = get_relativedelta(lead_max_str, 'H')
+    return lead_min, lead_max, no_max
+
+def _handle_lead_seq(config, lead_strings, lead_min=None, lead_max=None):
+    out_leads = []
+    leads = []
+    for lead in lead_strings:
+        relative_delta = get_relativedelta(lead, 'H')
+        if relative_delta is not None:
+            leads.append(relative_delta)
+        else:
+            config.logger.error(f'Invalid item {lead} in LEAD_SEQ. Exiting.')
+            return None
+
+    if lead_min is None and lead_max is None:
+        return leads
+
+    # add current time to leads to approximate month and year length
+    now_time = datetime.now()
+    lead_min_approx = now_time + lead_min
+    lead_max_approx = now_time + lead_max
+    for lead in leads:
+        lead_approx = now_time + lead
+        if lead_approx >= lead_min_approx and lead_approx <= lead_max_approx:
+            out_leads.append(lead)
+
+    return out_leads
+
+def _handle_init_seq(init_seq, input_dict, lead_min, lead_max):
+    out_leads = []
+    lead_min_hours = ti_get_hours_from_relativedelta(lead_min)
+    lead_max_hours = ti_get_hours_from_relativedelta(lead_max)
+
+    valid_hr = int(input_dict['valid'].strftime('%H'))
+    for init in init_seq:
+        if valid_hr >= init:
+            current_lead = valid_hr - init
+        else:
+            current_lead = valid_hr + (24 - init)
+
+        while current_lead <= lead_max_hours:
+            if current_lead >= lead_min_hours:
+                out_leads.append(get_relativedelta(current_lead, default_unit='H'))
+            current_lead += 24
+
+    out_leads = sorted(out_leads, key=lambda
+        rd: ti_get_seconds_from_relativedelta(rd, input_dict['valid']))
+    return out_leads
+
+def _handle_lead_groups(lead_groups):
+    """! Read groups of forecast leads and create a list with all unique items
+
+         @param lead_group dictionary where the values are lists of forecast
+         leads stored as relativedelta objects
+         @returns list of forecast leads stored as relativedelta objects
+    """
+    out_leads = []
+    for _, lead_seq in lead_groups.items():
+        for lead in lead_seq:
+            if lead not in out_leads:
+                out_leads.append(lead)
+
+    return out_leads
+
+def get_lead_sequence_groups(config):
+    # output will be a dictionary where the key will be the
+    #  label specified and the value will be the list of forecast leads
+    lead_seq_dict = {}
+    # used in plotting
+    all_conf = config.keys('config')
+    indices = []
+    regex = re.compile(r"LEAD_SEQ_(\d+)")
+    for conf in all_conf:
+        result = regex.match(conf)
+        if result is not None:
+            indices.append(result.group(1))
+
+    # loop over all possible variables and add them to list
+    for index in indices:
+        if config.has_option('config', f"LEAD_SEQ_{index}_LABEL"):
+            label = config.getstr('config', f"LEAD_SEQ_{index}_LABEL")
+        else:
+            log_msg = (f'Need to set LEAD_SEQ_{index}_LABEL to describe '
+                       f'LEAD_SEQ_{index}')
+            config.logger.error(log_msg)
+            return None
+
+        # get forecast list for n
+        lead_string_list = getlist(config.getstr('config', f'LEAD_SEQ_{index}'))
+        lead_seq = _handle_lead_seq(config,
+                                    lead_string_list,
+                                    lead_min=None,
+                                    lead_max=None)
+        # add to output dictionary
+        lead_seq_dict[label] = lead_seq
+
+    return lead_seq_dict
