+++ metplus/wrappers/gempak_to_cf_wrapper.py
@@ -12,8 +12,7 @@ Condition codes: 0 for success, 1 for failure
 
 import os
 
-from ..util import met_util as util
-from ..util import do_string_sub, skip_time
+from ..util import do_string_sub, skip_time, get_lead_sequence
 from ..util import time_util
 from . import CommandBuilder
 
@@ -75,7 +74,7 @@ class GempakToCFWrapper(CommandBuilder):
               Args:
                 @param input_dict dictionary containing timing information
         """
-        lead_seq = util.get_lead_sequence(self.config, input_dict)
+        lead_seq = get_lead_sequence(self.config, input_dict)
         for lead in lead_seq:
             self.clear()
             input_dict['lead'] = lead
