+++ metplus/wrappers/grid_diag_wrapper.py
@@ -12,11 +12,9 @@ Condition codes: 0 for success, 1 for failure
 
 import os
 
-from ..util import met_util as util
 from ..util import time_util
 from . import RuntimeFreqWrapper
-from ..util import do_string_sub
-from ..util import parse_var_list
+from ..util import do_string_sub, parse_var_list, sub_var_list
 
 '''!@namespace GridDiagWrapper
 @brief Wraps the Grid-Diag tool
@@ -187,7 +185,7 @@ class GridDiagWrapper(RuntimeFreqWrapper):
                 @param time_info time dictionary to use for string substitution
                 @returns True if field list could be built, False if not.
         """
-        field_list = util.sub_var_list(self.c_dict['VAR_LIST_TEMP'], time_info)
+        field_list = sub_var_list(self.c_dict['VAR_LIST_TEMP'], time_info)
         if not field_list:
             self.log_error("Could not get field information from config.")
             return False
