+++ metplus/wrappers/point_stat_wrapper.py
@@ -13,7 +13,6 @@ Condition codes: 0 for success, 1 for failure
 import os
 
 from ..util import getlistint
-from ..util import met_util as util
 from ..util import time_util
 from ..util import do_string_sub
 from . import CompareGriddedWrapper
